Removed some timestamp fields which didn't use them Factions are now stored in the database Routes' settings now save properly Added images for factions

+++ app/Http/Controllers/APIDungeonRouteController.php
@@ -20,31 +20,41 @@ class APIDungeonRouteController extends Controller
         /** @var DungeonRoute $dungeonroute */
         $dungeonroute = DungeonRoute::findOrNew($id);
 
+        if (!isset($dungeonroute->id)) {
+            $dungeonroute->author_id = \Auth::user()->id;
+        }
         $dungeonroute->dungeon_id = $request->get('dungeon', $dungeonroute->dungeon_id);
-        $dungeonroute->faction = $request->get('faction', $dungeonroute->faction);
+        $dungeonroute->faction_id = $request->get('faction', $dungeonroute->faction_id);
 
         // Update or insert it
         if ($dungeonroute->save()) {
-            // Remove old races
-            $dungeonroute->races()->delete();
+            $newRaces = $request->get('race', array());
 
-            // We don't _really_ care if this doesn't get saved properly, they can just set it again when editing.
-            foreach($request->get('race') as $key => $value){
-                $drpRace = new DungeonRoutePlayerRace();
-                $drpRace->index = $key;
-                $drpRace->race_id = $value;
-                $drpRace->dungeon_route_id = $dungeonroute->id;
-                $drpRace->save();
+            if (!empty($newRaces)) {
+                // Remove old races
+                $dungeonroute->races()->delete();
+
+                // We don't _really_ care if this doesn't get saved properly, they can just set it again when editing.
+                foreach ($newRaces as $key => $value) {
+                    $drpRace = new DungeonRoutePlayerRace();
+                    $drpRace->index = $key;
+                    $drpRace->race_id = $value;
+                    $drpRace->dungeon_route_id = $dungeonroute->id;
+                    $drpRace->save();
+                }
             }
 
-            // Remove old classes
-            $dungeonroute->classes()->delete();
-            foreach($request->get('class') as $key => $value){
-                $drpClass = new DungeonRoutePlayerClass();
-                $drpClass->index = $key;
-                $drpClass->class_id = $value;
-                $drpClass->dungeon_route_id = $dungeonroute->id;
-                $drpClass->save();
+            $newClasses = $request->get('class', array());
+            if (!empty($newClasses)) {
+                // Remove old classes
+                $dungeonroute->classes()->delete();
+                foreach ($newClasses as $key => $value) {
+                    $drpClass = new DungeonRoutePlayerClass();
+                    $drpClass->index = $key;
+                    $drpClass->class_id = $value;
+                    $drpClass->dungeon_route_id = $dungeonroute->id;
+                    $drpClass->save();
+                }
             }
         } else {
             abort(500, 'Unable to save dungeonroute');
